transfer Learning/requirement.txt
torch
numpy
pandas
joblib
matplotlib
scikit-learn
tqdm
hydra-core
torchvision
tensorboard
scipy
grad-cam
